hotfix: se reformo la estructura de los commits, en buscar solucion al problema de github actions
