+++ scripts/cb-test.sh
@@ -212,3 +212,5 @@ if [ -n "$TODOS_CB_URL" ]; then
 fi
 
 echo "== Listo =="
+
+
